# Add top-level schema caches to Schema::Visibility for better performance