--- a/lib/graphql/schema/visibility.rb
+++ b/lib/graphql/schema/visibility.rb
@@ -13,6 +13,9 @@ module GraphQL
       # @param migration_errors [Boolean] if `true`, raise an error when `Visibility` and `Warden` return different results
       def self.use(schema, dynamic: false, profiles: EmptyObjects::EMPTY_HASH, preload: (defined?(Rails) ? Rails.env.production? : nil), migration_errors: false)
         schema.visibility = self.new(schema, dynamic: dynamic, preload: preload, profiles: profiles, migration_errors: migration_errors)
+        if preload
+          schema.visibility.preload
+        end
       end
 
       def initialize(schema, dynamic:, preload:, profiles:, migration_errors:)
@@ -26,11 +29,63 @@ module GraphQL
         @cached_profiles = {}
         @dynamic = dynamic
         @migration_errors = migration_errors
-        if preload
-          # Traverse the schema now (and in the *_configured hooks below)
-          # To make sure things are loaded during boot
-          @preloaded_types = Set.new
-          types_to_visit = [
+        @top_level = TopLevel.new(@schema)
+      end
+
+      class TopLevel
+        def initialize(schema)
+          @schema = schema
+          @loaded_all = false
+          @interface_type_memberships = Hash.new { |h, interface_type| h[interface_type] = [] }.compare_by_identity
+          @directives = []
+        end
+
+        def directives
+          load_all
+          @directives
+        end
+
+        def interface_type_memberships
+          load_all
+          @interface_type_memberships
+        end
+
+        private
+
+        def load_all
+          # TODO thread-safe
+          @loaded_all ||= begin
+            Visit.each(@schema) do |member|
+              if member.is_a?(Class)
+                if member < GraphQL::Schema::Directive
+                  @directives << member
+                elsif member.respond_to?(:interface_type_memberships)
+                  member.interface_type_memberships.each do |itm|
+                    @interface_type_memberships[itm.abstract_type] << itm
+                  end
+                end
+              end
+              true
+            end
+            true
+          end
+        end
+      end
+
+      class Visit
+        def self.each(schema, &block)
+          self.new(schema).visit_each(&block)
+        end
+
+        def initialize(schema)
+          @schema = schema
+        end
+
+        def visit_each
+          visited_types = Set.new.compare_by_identity
+          late_union_type_memberships = []
+          visited_directives = Set.new.compare_by_identity
+          unvisited_types = [
             @schema.query,
             @schema.mutation,
             @schema.subscription,
@@ -38,15 +93,119 @@ module GraphQL
             *@schema.introspection_system.entry_points.map { |ep| ep.type.unwrap },
             *@schema.orphan_types,
           ]
-          # Root types may have been nil:
-          types_to_visit.compact!
-          ensure_all_loaded(types_to_visit)
 
-          profiles.each do |profile_name, example_ctx|
-            example_ctx[:visibility_profile] = profile_name
-            prof = profile_for(example_ctx, profile_name)
-            prof.all_types # force loading
+          possible_types_to_add = []
+          directives_to_visit = []
+
+          unvisited_types.compact!
+
+          @schema.directives.each_value { |dir_class| yield(dir_class) }
+
+          while unvisited_types.any? || late_union_type_memberships.any?
+            while (type = unvisited_types.pop)
+              if visited_types.add?(type) && yield(type)
+                directives_to_visit.concat(type.directives)
+                append_next_visits(type, unvisited_types, possible_types_to_add, late_union_type_memberships, directives_to_visit)
+              end
+            end
+
+            directives_to_visit.each do |dir|
+              dir_class = dir.class
+              if visited_directives.add?(dir_class)
+                yield(dir_class)
+              end
+            end
+
+            # possible_types_to_add.each do |int_t|
+            #   int_t_mems = @interface_type_memberships[int_t]
+            #   int_t_mems.each do |int_t_mem|
+            #     unvisited_types << int_t_mem.object_type
+            #   end
+            # end
+            # possible_types_to_add.clear
+
+            while (union_tm = late_union_type_memberships.shift)
+              late_obj_t = union_tm.object_type
+              obj_t = all_types[late_obj_t.graphql_name] || raise("Failed to resolve union type membership: #{late_obj_t.graphql_name.inspect} on #{union_tm.inspect}")
+              union_tm.abstract_type.assign_type_membership_object_type(obj_t)
+            end
           end
+          nil
+        end
+
+        def append_next_visits(type, unvisited_types, possible_types_to_add, late_union_type_memberships, directives)
+          case type.kind.name
+          when "OBJECT", "INTERFACE"
+            type.interface_type_memberships.each do |itm|
+              unvisited_types << itm.abstract_type
+            end
+            if type.kind.interface?
+              unvisited_types.concat(type.orphan_types)
+            end
+
+            type.all_field_definitions.each do |field|
+              field.ensure_loaded
+              directives.concat(field.directives)
+              field_type = field.type.unwrap
+              if field_type.kind.interface?
+                possible_types_to_add << field_type
+              end
+              unvisited_types << field_type
+              field.all_argument_definitions.each do |argument|
+                directives.concat(argument.directives)
+                unvisited_types << argument.type.unwrap
+              end
+            end
+          when "INPUT_OBJECT"
+            type.all_argument_definitions.each do |argument|
+              directives.concat(argument.directives)
+              unvisited_types << argument.type.unwrap
+            end
+          when "UNION"
+            type.type_memberships.each do |tm|
+              obj_t = tm.object_type
+              if obj_t.is_a?(GraphQL::Schema::LateBoundType)
+                late_union_type_memberships << tm
+              else
+                if obj_t.is_a?(String)
+                  obj_t = Member::BuildType.constantize(obj_t)
+                  tm.object_type = obj_t
+                end
+                unvisited_types << obj_t
+              end
+            end
+          when "ENUM"
+            type.all_enum_value_definitions.each do |val|
+              directives.concat(val.directives)
+            end
+          when "SCALAR"
+            # pass
+          else
+            raise "Invariant: unhandled type kind: #{type.kind.inspect}"
+          end
+        end
+      end
+
+      def preload
+        # Traverse the schema now (and in the *_configured hooks below)
+        # To make sure things are loaded during boot
+        @preloaded_types = Set.new
+        types_to_visit = [
+          @schema.query,
+          @schema.mutation,
+          @schema.subscription,
+          *@schema.introspection_system.types.values,
+          *@schema.introspection_system.entry_points.map { |ep| ep.type.unwrap },
+          *@schema.orphan_types,
+        ]
+        # Root types may have been nil:
+        types_to_visit.compact!
+        ensure_all_loaded(types_to_visit)
+
+        @profiles.each do |profile_name, example_ctx|
+          example_ctx[:visibility_profile] = profile_name
+          prof = profile_for(example_ctx, profile_name)
+          prof.all_types # force loading
         end
       end
 
@@ -132,7 +291,10 @@ module GraphQL
         end
       end
 
-      private
+      attr_reader :top_level
+
+      # @api private
+      attr_reader :unfiltered_interface_type_memberships
 
       def top_level_profile(refresh: false)
         if refresh
@@ -140,6 +302,8 @@ module GraphQL
         end
         @top_level_profile ||= @schema.visibility_profile_class.new(context: Query::NullContext.instance, schema: @schema)
       end
+
+      private
 
       def ensure_all_loaded(types_to_visit)
         while (type = types_to_visit.shift)