… lib/graphql/schema/visibility/profile.rb
… lib/graphql/schema/visibility/visit.rb
… lib/graphql/static_validation/validation_context.rb
… lib/graphql/testing/helpers.rb
… spec/graphql/schema/visibility/profile_spec.rb
… spec/graphql/schema/visibility_spec.rb